🚚 Move FileActionButtons component to /documents/components
